feat: canonical ExecutionReport, OnError contract, and cross-SDK API stabilisation
Runtime
- Add ExecutionReport { ok, results, succeeded, failed } — stable return
  shape for every execute call regardless of input type.
- Add OnError::Stop / Continue to replace hard-coded stop_on_error.
- Rewrite execute_batch_nodes_inner with BatchKey-based grouping and
  flush barriers so statement ordering is preserved and side-effecting
  prepare (UPSERT auto-creation) no longer runs ahead of explicit DDL.

Core parser
- Remove try_parse() — it duplicated parse() without being lighter.

SDK API normalisation (Python / Node / WASM)
- parse() always returns a list (single or semicolon-delimited script).
- Remove parse_batch / parseBatch / parseBatchJson from all SDKs.
- Add on_error param to Python and Node execute methods.
- Add structured QqlError serialisation (code / kind / span) at the
  N-API boundary; JS buildError() extracts and attaches the fields.
- WASM: embedded TypeScript types via typescript_custom_section;
  canonical ExecResponse shape with data: null; batch cardinality
  checks; WasmOnError parity with runtime.
- NQQL: platform-specific native binary loading (musl/gnu/darwin/win32),
  Client wrapper class, validateOptions, compileQuery export.
- PyQQL: HttpEmbedder validation, Stmt + execute_async exports.

CI
- Platform jobs now build and test actual SDK artifacts:
  pyqql (pip install + unittest), nqql (npm install + npm test +
  npm pack --dry-run), wasm (wasm-pack build).
- Set PYO3_USE_ABI3_FORWARD_COMPATIBILITY for Python 3.14.

Docs / skills / examples / benchmarks — updated for new API shapes.

+++ crates/qql-cli/src/output.rs
@@ -10,8 +10,8 @@ pub struct ScriptResponse {
     pub ok: bool,
     pub command: String,
     pub path: String,
-    pub succeeded: u32,
-    pub failed: u32,
+    pub succeeded: usize,
+    pub failed: usize,
     pub message: String,
 }
